Created features.py
This file contains two methods for the moment. They allow to compute some possible features. To be tested when models function.
The first method takes an image and a string to specify which filter to apply to the image. It returns the filtered image.
The second method takes an image and possibly a factor to adjust the contrast of the image.